- Improved base damage of players - Implemented 2 bulkan skills (SwordMastery and AxeMastery) - Fixed the loading of some player stats when the game loads - Added a new GM command (@skilllevel) - Fixed other minor bugs
git-svn-id: http://reunion.googlecode.com/svn/trunk@200 d5b2e400-7aba-11de-9499-4f8beaa274e3

+++ jreunion/src/com/googlecode/reunion/jreunion/game/HumanPlayer.java
@@ -89,7 +89,7 @@ public class HumanPlayer extends Player {
 	}
 
 	@Override
-	public int getBaseDamage() {
+	public float getBaseDamage() {
 		return (getLevel() / 6) + (getDexterity() / 4)+ getStrength();
 		
 	}
